pyproject.toml
[project]
name = "linkedin-etl-project"
version = "0.1.0"
description = ""
authors = [
    {name = "nikki xie",email = "nikki099@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "snowflake-connector-python (>=3.15.0,<4.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "snowflake-sqlalchemy (>=1.7.3,<2.0.0)",
    "dbt-core (>=1.9.4,<2.0.0)",
    "dbt-snowflake (>=1.9.3,<2.0.0)",
    "googletrans (==4.0.0rc1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
